% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscognostics.R
\name{tsmeasures}
\alias{tsmeasures}
\title{Compute a collection of cognostics for multivariate time series.}
\usage{
tsmeasures(y, normalise = TRUE, width)
}
\arguments{
\item{y}{a \code{ts} object. A multivariate time series.}

\item{normalise}{logical. If TRUE, scaling time series to be normally distributed
with mean 0 and 1.}

\item{width}{integer. A window size specified for variance change, level shift,
and lumpiness.}
}
\value{
Cognostics matrix for every time series.
}
\description{
The function computes a set of cognostics to summarize statistical features
of multivariate time series.
}

